feat: move data-driven to a independent module
