+++ modules/data-driven/src/contentTesting/java/test/kasuga/data_driven/DataDrivenTestCreativeTab.java
@@ -1,4 +1,4 @@
-package test.kasuga.core.data_driven;
+package test.kasuga.data_driven;
 
 import io.micronaut.context.annotation.Context;
 import lib.kasuga.registration.minecraft.creative_tab.CreativeTabReg;
