+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/blockstates/be_test_block.json
