+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/blockstates/detail_slab.json
