+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/blockstates/inherited_block.json
