+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/blockstates/overridden_block.json
