+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/blockstates/simple_panel.json
