+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/blockstates/standalone_block.json
