+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/block/be_test_block.json
