+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/block/detail_slab.json
