+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/block/inherited_block.json
