+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/block/simple_panel.json
