+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/block/standalone_block.json
