+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/item/be_test_block.json
