+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/item/detail_slab.json
