+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/item/inherited_block.json
