+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/item/overridden_block.json
