+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/item/simple_panel.json
