+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/item/standalone_block.json
