+++ modules/data-driven/src/contentTesting/resources/assets/kasuga_lib/models/item/test_material.json
