+++ modules/data-driven/src/contentTesting/resources/data/kasuga_lib/kasugalib/inheritance.json
