+++ modules/data-driven/src/main/java/lib/kasuga/registration/data_driven/handler/ItemDef.java
