+++ modules/data-driven/src/test/java/lib/kasuga/test/registration/data_driven/FactoryRegistryTest.java
