+++ settings.gradle
@@ -13,6 +13,7 @@ rootProject.name = 'KasugaLibNext'
 includeBuild './modules/gradle-plugin'
 include ':modules:utils'
 include ':modules:core'
+include ':modules:data-driven'
 include ':modules:inject'
 include ':modules:create'
 include ':modules:scripting'
